Api/requirements.txt
kafka-python-ng
pyspark == 3.2.4
pymongo
pandas